requirements.txt
amqp==5.3.1
anyio==4.12.1
beautifulsoup4==4.14.3
billiard==4.2.4
bonbast==1.0.2
bs4==0.0.2
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
redis==7.2.0
requests==2.32.5
rich==14.3.2
six==1.17.0
socksio==1.0.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
yfinance==0.2.36
pytz==2024.1
lxml==5.1.0
